+++ drafthelper-app/src/index.py
@@ -1,3 +1,9 @@
-from services.interface import App
-# TÄNNE TULEE MYÖHEMMIN GRAAFISEN LIITTYMÄN KÄYNNISTYS
-App()
+from tkinter import Tk
+from interface.ui import UI
+
+window = Tk()
+window.title('Draft Helper - Fantasy draft simulator')
+
+ui = UI(window)
+ui.show_settings_view()
+window.mainloop()