+++ drafthelper-app/src/index_txt.py
@@ -1,0 +1,4 @@
+from services.interface import App
+
+interface = App()
+interface.start_txt()